regression 1st commit

+++ src/gaknn/predictor/Predictor1.java
@@ -36,7 +36,7 @@ public class Predictor1 extends Predictor {
         
         for (int i=0; i<m_K; i++){
             int index = ((Pair)simList[i]).Index();
-            ClassIndex = trainSet[index].GetClassIndex();
+            ClassIndex = (int)trainSet[index].GetClassIndex();
             vote[ClassIndex]+= ((Pair)simList[i]).Value();
         }
         
@@ -68,7 +68,7 @@ public class Predictor1 extends Predictor {
         
         for (int i=0; i<m_K; i++){
             int index = ((Pair)simList[i]).Index();
-            ClassIndex = trainSet[index].GetClassIndex();
+            ClassIndex = (int)trainSet[index].GetClassIndex();
             vote[ClassIndex]+= ((Pair)simList[i]).Value();
         }
